+++ sdk/websites/Azure.ResourceManager.AppService/src/Customization/Models/AppCertificateData.Serialization.cs
@@ -33,7 +33,7 @@ namespace Azure.ResourceManager.AppService
             Optional<string> issuer = default;
             Optional<DateTimeOffset> issueDate = default;
             Optional<DateTimeOffset> expirationDate = default;
-            Optional<BinaryData> thumbprint = default;
+            Optional<string> thumbprintString = default;
             Optional<bool> valid = default;
             Optional<byte[]> cerBlob = default;
             Optional<string> publicKeyHash = default;
@@ -187,7 +187,7 @@ namespace Azure.ResourceManager.AppService
                                 property0.ThrowNonNullablePropertyIsNull();
                                 continue;
                             }
-                            thumbprint = BinaryData.FromString(property0.Value.GetRawText());
+                            thumbprintString = property0.Value.GetString();
                             continue;
                         }
                         if (property0.NameEquals("valid"u8))
@@ -274,7 +274,7 @@ namespace Azure.ResourceManager.AppService
                     continue;
                 }
             }
-            return new AppCertificateData(id, name, type, systemData.Value, Optional.ToDictionary(tags), location, password.Value, friendlyName.Value, subjectName.Value, Optional.ToList(hostNames), pfxBlob.Value, siteName.Value, selfLink.Value, issuer.Value, Optional.ToNullable(issueDate), Optional.ToNullable(expirationDate), thumbprint.Value, Optional.ToNullable(valid), cerBlob.Value, publicKeyHash.Value, hostingEnvironmentProfile.Value, keyVaultId.Value, keyVaultSecretName.Value, Optional.ToNullable(keyVaultSecretStatus), serverFarmId.Value, canonicalName.Value, domainValidationMethod.Value, kind.Value);
+            return new AppCertificateData(id, name, type, systemData.Value, Optional.ToDictionary(tags), location, password.Value, friendlyName.Value, subjectName.Value, Optional.ToList(hostNames), pfxBlob.Value, siteName.Value, selfLink.Value, issuer.Value, Optional.ToNullable(issueDate), Optional.ToNullable(expirationDate), thumbprintString, Optional.ToNullable(valid), cerBlob.Value, publicKeyHash.Value, hostingEnvironmentProfile.Value, keyVaultId.Value, keyVaultSecretName.Value, Optional.ToNullable(keyVaultSecretStatus), serverFarmId.Value, canonicalName.Value, domainValidationMethod.Value, kind.Value);
         }
     }
 }